Everything's on fire
All the obvious errors are gone, mostly because I cut loads of corners.
The only thing that really survived this "port" is the nice spiffy edit
menu, and even then it doesn't work right; the game will crash if you
attempt to make a new chart (presumably, it has a similar reaction to
trying to delete a chart).

I don't like the shape of things... at all. But it'll have to do.